refactor: move prisma.ts file to lib folder
